backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
asyncpg==0.29.0
redis[asyncio]==5.0.8
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic-settings==2.3.0
pydantic[email]>=2.3.0
httpx==0.27.0
celery[redis]==5.4.0
cachetools==5.3.3
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk[fastapi]==2.7.0
structlog==24.2.0
slowapi==0.1.9
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
anyio==4.4.0
fakeredis==2.23.2
